security/pyproject.toml
[project]
name = "security"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
entrypoint = "main:app"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.136.0",
    "pyjwt>=2.12.1",
]
